requirements.txt
Click==7.0
dlib==19.17.0
face-recognition==1.2.3
face-recognition-models==0.3.0
mysql-connector-python==8.0.17
nltk==3.4.5
numpy==1.17.2
opencv-python==4.1.1.26
Pillow==6.1.0
protobuf==3.9.1
PyAudio==0.2.11
pypiwin32==223
pyttsx3==2.71
pywin32==225
six==1.12.0
SpeechRecognition==3.8.1
textblob==0.15.3
